#152 Change place defaultChannel method
