+++ app/src/androidTest/java/ch/epfl/sweng/zuluzulu/PostFragmentTest.java
@@ -8,7 +8,6 @@ import org.junit.runner.RunWith;
 
 import ch.epfl.sweng.zuluzulu.Fragments.PostFragment;
 import ch.epfl.sweng.zuluzulu.Fragments.SuperFragment;
-import ch.epfl.sweng.zuluzulu.Structure.Utils;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
@@ -20,7 +19,7 @@ public class PostFragmentTest extends TestWithLogin {
 
     @Before
     public void init() {
-        SuperFragment fragment = PostFragment.newInstance(getUser(), Utils.defaultChannel());
+        SuperFragment fragment = PostFragment.newInstance(getUser(), Utility.defaultChannel());
         mActivityRule.getActivity().openFragment(fragment);
     }
